+++ src/main/java/com/structurizr/dsl/StructurizrDslParser.java
@@ -31,6 +31,7 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
 
     private IdentifierScope identifierScope = IdentifierScope.Flat;
     private Stack<DslContext> contextStack;
+    private Set<String> parsedTokens = new HashSet<>();
     private IdentifiersRegister identifiersRegister;
     private Map<String, Constant> constants;
 
@@ -47,7 +48,7 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
      */
     public StructurizrDslParser(StructurizrDslParserListener _parserListener) {
         contextStack = new Stack<>();
-        identifersRegister = new IdentifiersRegister();
+        identifiersRegister = new IdentifiersRegister();
         constants = new HashMap<>();
         parserListener = _parserListener;
     }
@@ -217,7 +218,7 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
                         // do nothing
 
                     } else if (DslContext.CONTEXT_END_TOKEN.equals(tokens.get(0))) {
-                        parserListener.onEndContext(file, lineNumber, contextStack.peek().getClass().getSimpleName());
+                        parserListener.onEndContext(dslFile, lineNumber, contextStack.peek().getClass().getSimpleName());
                         endContext();
 
                     } else if (tokens.size() > 2 && RELATIONSHIP_TOKEN.equals(tokens.get(1)) && (inContext(ModelDslContext.class) || inContext(EnterpriseDslContext.class) || inContext(CustomElementDslContext.class) || inContext(PersonDslContext.class) || inContext(SoftwareSystemDslContext.class) || inContext(ContainerDslContext.class) || inContext(ComponentDslContext.class) || inContext(DeploymentEnvironmentDslContext.class) || inContext(DeploymentNodeDslContext.class) || inContext(InfrastructureNodeDslContext.class) || inContext(SoftwareSystemInstanceDslContext.class) || inContext(ContainerInstanceDslContext.class))) {
@@ -228,7 +229,7 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
                         }
 
                         registerIdentifier(identifier, relationship);
-                        parserListener.onParsedRelationShip(file, lineNumber, identifier, relationship);
+                        parserListener.onParsedRelationShip(dslFile, lineNumber, identifier, relationship);
 
                     } else if (tokens.size() >= 2 && RELATIONSHIP_TOKEN.equals(tokens.get(0)) && (inContext(CustomElementDslContext.class) || inContext(PersonDslContext.class) || inContext(SoftwareSystemDslContext.class) || inContext(ContainerDslContext.class) || inContext(ComponentDslContext.class) || inContext(DeploymentNodeDslContext.class) || inContext(InfrastructureNodeDslContext.class) || inContext(SoftwareSystemInstanceDslContext.class) || inContext(ContainerInstanceDslContext.class))) {
                         Relationship relationship = new ImplicitRelationshipParser().parse(getContext(ModelItemDslContext.class), tokens.withoutContextStartToken());
@@ -238,7 +239,7 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
                         }
 
                         registerIdentifier(identifier, relationship);
-                        parserListener.onParsedRelationShip(file, lineNumber, identifier, relationship);
+                        parserListener.onParsedRelationShip(dslFile, lineNumber, identifier, relationship);
 
                     } else if (REF_TOKEN.equalsIgnoreCase(firstToken) && (inContext(ModelDslContext.class))) {
                         ModelItem modelItem = new RefParser().parse(getContext(), tokens.withoutContextStartToken());
@@ -284,7 +285,7 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
 
                         registerIdentifier(identifier, customElement);
                         
-                        parserListener.onParsedModelElement(file, lineNumber, identifier, customElement);
+                        parserListener.onParsedModelElement(dslFile, lineNumber, identifier, customElement);
 
 
                     } else if (PERSON_TOKEN.equalsIgnoreCase(firstToken) && (inContext(ModelDslContext.class) || inContext(EnterpriseDslContext.class))) {
@@ -296,7 +297,7 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
 
                         registerIdentifier(identifier, person);
                         
-                        parserListener.onParsedModelElement(file, lineNumber, identifier, person);
+                        parserListener.onParsedModelElement(dslFile, lineNumber, identifier, person);
 
                     } else if (SOFTWARE_SYSTEM_TOKEN.equalsIgnoreCase(firstToken) && (inContext(ModelDslContext.class) || inContext(EnterpriseDslContext.class))) {
                         SoftwareSystem softwareSystem = new SoftwareSystemParser().parse(getContext(GroupableDslContext.class), tokens.withoutContextStartToken());
@@ -307,7 +308,7 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
 
                         registerIdentifier(identifier, softwareSystem);
 
-                        parserListener.onParsedModelElement(file, lineNumber, identifier, softwareSystem);
+                        parserListener.onParsedModelElement(dslFile, lineNumber, identifier, softwareSystem);
 
                     } else if (CONTAINER_TOKEN.equalsIgnoreCase(firstToken) && inContext(SoftwareSystemDslContext.class)) {
                         Container container = new ContainerParser().parse(getContext(SoftwareSystemDslContext.class), tokens.withoutContextStartToken());
@@ -318,7 +319,7 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
 
                         registerIdentifier(identifier, container);
                         
-                        parserListener.onParsedModelElement(file, lineNumber, identifier, container);
+                        parserListener.onParsedModelElement(dslFile, lineNumber, identifier, container);
 
                     } else if (COMPONENT_TOKEN.equalsIgnoreCase(firstToken) && inContext(ContainerDslContext.class)) {
                         Component component = new ComponentParser().parse(getContext(ContainerDslContext.class), tokens.withoutContextStartToken());
@@ -329,21 +330,21 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
 
                         registerIdentifier(identifier, component);
                         
-                        parserListener.onParsedModelElement(file, lineNumber, identifier, component);
+                        parserListener.onParsedModelElement(dslFile, lineNumber, identifier, component);
 
                     } else if (GROUP_TOKEN.equalsIgnoreCase(firstToken) && inContext(ModelDslContext.class) && !getContext(ModelDslContext.class).hasGroup()) {
                         ElementGroup group = new GroupParser().parse(tokens.withoutContextStartToken());
 
                         startContext(new ModelDslContext(group));
                         registerIdentifier(identifier, group);
-                        parserListener.onParsedModelElement(file, lineNumber, identifier, group);
+                        parserListener.onParsedModelElement(dslFile, lineNumber, identifier, group);
 
                     } else if (GROUP_TOKEN.equalsIgnoreCase(firstToken) && inContext(EnterpriseDslContext.class) && !getContext(EnterpriseDslContext.class).hasGroup()) {
                         ElementGroup group = new GroupParser().parse(tokens.withoutContextStartToken());
 
                         startContext(new EnterpriseDslContext(group));
                         registerIdentifier(identifier, group);
-                        parserListener.onParsedModelElement(file, lineNumber, identifier, group);
+                        parserListener.onParsedModelElement(dslFile, lineNumber, identifier, group);
 
                     } else if (GROUP_TOKEN.equalsIgnoreCase(firstToken) && inContext(SoftwareSystemDslContext.class) && !getContext(SoftwareSystemDslContext.class).hasGroup()) {
                         ElementGroup group = new GroupParser().parse(tokens.withoutContextStartToken());
@@ -352,7 +353,7 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
                         group.setParent(softwareSystem);
                         startContext(new SoftwareSystemDslContext(softwareSystem, group));
                         registerIdentifier(identifier, group);
-                        parserListener.onParsedModelElement(file, lineNumber, identifier, group);
+                        parserListener.onParsedModelElement(dslFile, lineNumber, identifier, group);
 
                     } else if (GROUP_TOKEN.equalsIgnoreCase(firstToken) && inContext(ContainerDslContext.class) && !getContext(ContainerDslContext.class).hasGroup()) {
                         ElementGroup group = new GroupParser().parse(tokens.withoutContextStartToken());
@@ -361,18 +362,18 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
                         group.setParent(container);
                         startContext(new ContainerDslContext(container, group));
                         registerIdentifier(identifier, group);
-                        parserListener.onParsedModelElement(file, lineNumber, identifier, group);
+                        parserListener.onParsedModelElement(dslFile, lineNumber, identifier, group);
 
-                    } else if (TAGS_TOKEN.equalsIgnoreCase(firstToken) && inContext(ModelItemDslContext.class)) {
+                    } else if (TAGS_TOKEN.equalsIgnoreCase(firstToken) && inContext(ModelItemDslContext.class) && !getContext(ModelItemDslContext.class).hasGroup()) {
                         new ModelItemParser().parseTags(getContext(ModelItemDslContext.class), tokens);
 
-                    } else if (DESCRIPTION_TOKEN.equalsIgnoreCase(firstToken) && inContext(ModelItemDslContext.class) && getContext(ModelItemDslContext.class).getModelItem() instanceof Element) {
+                    } else if (DESCRIPTION_TOKEN.equalsIgnoreCase(firstToken) && inContext(ModelItemDslContext.class) && getContext(ModelItemDslContext.class).getModelItem() instanceof Element && !getContext(ModelItemDslContext.class).hasGroup()) {
                         new ModelItemParser().parseDescription(getContext(ModelItemDslContext.class), tokens);
 
-                    } else if (TECHNOLOGY_TOKEN.equalsIgnoreCase(firstToken) && inContext(ContainerDslContext.class)) {
+                    } else if (TECHNOLOGY_TOKEN.equalsIgnoreCase(firstToken) && inContext(ContainerDslContext.class) && !getContext(ContainerDslContext.class).hasGroup()) {
                         new ContainerParser().parseTechnology(getContext(ContainerDslContext.class), tokens);
 
-                    } else if (TECHNOLOGY_TOKEN.equalsIgnoreCase(firstToken) && inContext(ComponentDslContext.class)) {
+                    } else if (TECHNOLOGY_TOKEN.equalsIgnoreCase(firstToken) && inContext(ComponentDslContext.class) && !getContext(ComponentDslContext.class).hasGroup()) {
                         new ComponentParser().parseTechnology(getContext(ComponentDslContext.class), tokens);
 
                     } else if (TECHNOLOGY_TOKEN.equalsIgnoreCase(firstToken) && inContext(DeploymentNodeDslContext.class)) {
@@ -381,10 +382,16 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
                     } else if (TECHNOLOGY_TOKEN.equalsIgnoreCase(firstToken) && inContext(InfrastructureNodeDslContext.class)) {
                         new InfrastructureNodeParser().parseTechnology(getContext(InfrastructureNodeDslContext.class), tokens);
 
-                    } else if (URL_TOKEN.equalsIgnoreCase(firstToken) && inContext(ModelItemDslContext.class)) {
+                    } else if (INSTANCES_TOKEN.equalsIgnoreCase(firstToken) && inContext(DeploymentNodeDslContext.class)) {
+                        new DeploymentNodeParser().parseInstances(getContext(DeploymentNodeDslContext.class), tokens);
+
+                    } else if (URL_TOKEN.equalsIgnoreCase(firstToken) && inContext(ModelItemDslContext.class) && !getContext(ModelItemDslContext.class).hasGroup()) {
                         new ModelItemParser().parseUrl(getContext(ModelItemDslContext.class), tokens);
 
-                    } else if (PROPERTIES_TOKEN.equalsIgnoreCase(firstToken) && inContext(ModelItemDslContext.class)) {
+                    } else if (PROPERTIES_TOKEN.equalsIgnoreCase(firstToken) && inContext(ConfigurationDslContext.class) && !getContext(ModelItemDslContext.class).hasGroup()) {
+                        startContext(new PropertiesDslContext(getContext(ConfigurationDslContext.class).getWorkspace()));
+
+                    } else if (PROPERTIES_TOKEN.equalsIgnoreCase(firstToken) && inContext(ModelItemDslContext.class) && !getContext(ModelItemDslContext.class).hasGroup()) {
                         startContext(new PropertiesDslContext(getContext(ModelItemDslContext.class).getModelItem()));
 
                     } else if (PROPERTIES_TOKEN.equalsIgnoreCase(firstToken) && inContext(ViewsDslContext.class)) {
@@ -393,19 +400,23 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
                     } else if (inContext(PropertiesDslContext.class)) {
                         new PropertyParser().parse(getContext(PropertiesDslContext.class), tokens);
 
-                    } else if (PERSPECTIVES_TOKEN.equalsIgnoreCase(firstToken) && inContext(ModelItemDslContext.class)) {
+                    } else if (PERSPECTIVES_TOKEN.equalsIgnoreCase(firstToken) && inContext(ModelItemDslContext.class) && !getContext(ModelItemDslContext.class).hasGroup()) {
                         startContext(new ModelItemPerspectivesDslContext(getContext(ModelItemDslContext.class).getModelItem()));
 
                     } else if (inContext(ModelItemPerspectivesDslContext.class)) {
                         new ModelItemParser().parsePerspective(getContext(ModelItemPerspectivesDslContext.class), tokens);
 
                     } else if (WORKSPACE_TOKEN.equalsIgnoreCase(firstToken) && contextStack.empty()) {
+                        if (parsedTokens.contains(WORKSPACE_TOKEN)) {
+                            throw new RuntimeException("Multiple workspaces are not permitted in a DSL definition");
+                        }
                         DslParserContext dslParserContext = new DslParserContext(dslFile, restricted);
                         dslParserContext.setIdentifierRegister(identifiersRegister);
 
                         workspace = new WorkspaceParser().parse(dslParserContext, tokens.withoutContextStartToken());
                         extendingWorkspace = !workspace.getModel().isEmpty();
                         startContext(new WorkspaceDslContext());
+                        parsedTokens.add(WORKSPACE_TOKEN);
                     } else if (IMPLIED_RELATIONSHIPS_TOKEN.equalsIgnoreCase(firstToken) || IMPLIED_RELATIONSHIPS_TOKEN.substring(1).equalsIgnoreCase(firstToken)) {
                         new ImpliedRelationshipsParser().parse(getContext(), tokens);
 
@@ -416,10 +427,20 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
                         new WorkspaceParser().parseDescription(getContext(), tokens);
 
                     } else if (MODEL_TOKEN.equalsIgnoreCase(firstToken) && inContext(WorkspaceDslContext.class)) {
+                        if (parsedTokens.contains(MODEL_TOKEN)) {
+                            throw new RuntimeException("Multiple models are not permitted in a DSL definition");
+                        }
+
                         startContext(new ModelDslContext());
+                        parsedTokens.add(MODEL_TOKEN);
 
                     } else if (VIEWS_TOKEN.equalsIgnoreCase(firstToken) && inContext(WorkspaceDslContext.class)) {
+                        if (parsedTokens.contains(VIEWS_TOKEN)) {
+                            throw new RuntimeException("Multiple view sets are not permitted in a DSL definition");
+                        }
+
                         startContext(new ViewsDslContext());
+                        parsedTokens.add(VIEWS_TOKEN);
 
                     } else if (BRANDING_TOKEN.equalsIgnoreCase(firstToken) && inContext(ViewsDslContext.class)) {
                         startContext(new BrandingDslContext(dslFile));
@@ -439,15 +460,15 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
 
                     } else if (ELEMENT_STYLE_BACKGROUND_TOKEN.equalsIgnoreCase(firstToken) && inContext(ElementStyleDslContext.class)) {
                         new ElementStyleParser().parseBackground(getContext(ElementStyleDslContext.class), tokens);
-                        parserListener.onParsedColor(file, lineNumber);
+                        parserListener.onParsedColor(dslFile, lineNumber);
                         
                     } else if ((ELEMENT_STYLE_COLOUR_TOKEN.equalsIgnoreCase(firstToken) || ELEMENT_STYLE_COLOR_TOKEN.equalsIgnoreCase(firstToken)) && inContext(ElementStyleDslContext.class)) {
                         new ElementStyleParser().parseColour(getContext(ElementStyleDslContext.class), tokens);
-                        parserListener.onParsedColor(file, lineNumber);
+                        parserListener.onParsedColor(dslFile, lineNumber);
 
                     } else if (ELEMENT_STYLE_STROKE_TOKEN.equalsIgnoreCase(firstToken) && inContext(ElementStyleDslContext.class)) {
                         new ElementStyleParser().parseStroke(getContext(ElementStyleDslContext.class), tokens);
-                        parserListener.onParsedColor(file, lineNumber);
+                        parserListener.onParsedColor(dslFile, lineNumber);
 
                     } else if (ELEMENT_STYLE_SHAPE_TOKEN.equalsIgnoreCase(firstToken) && inContext(ElementStyleDslContext.class)) {
                         new ElementStyleParser().parseShape(getContext(ElementStyleDslContext.class), tokens);
@@ -485,7 +506,7 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
 
                     } else if ((RELATIONSHIP_STYLE_COLOUR_TOKEN.equalsIgnoreCase(firstToken) || RELATIONSHIP_STYLE_COLOR_TOKEN.equalsIgnoreCase(firstToken)) && inContext(RelationshipStyleDslContext.class)) {
                         new RelationshipStyleParser().parseColour(getContext(RelationshipStyleDslContext.class), tokens);
-                        parserListener.onParsedColor(file, lineNumber);
+                        parserListener.onParsedColor(dslFile, lineNumber);
 
                     } else if (RELATIONSHIP_STYLE_DASHED_TOKEN.equalsIgnoreCase(firstToken) && inContext(RelationshipStyleDslContext.class)) {
                         new RelationshipStyleParser().parseDashed(getContext(RelationshipStyleDslContext.class), tokens);
@@ -531,7 +552,7 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
                         }
 
                         registerIdentifier(identifier, deploymentNode);
-                        parserListener.onParsedModelElement(file, lineNumber, identifier, deploymentNode);
+                        parserListener.onParsedModelElement(dslFile, lineNumber, identifier, deploymentNode);
                         
                     } else if (INFRASTRUCTURE_NODE_TOKEN.equalsIgnoreCase(firstToken) && inContext(DeploymentNodeDslContext.class)) {
                         InfrastructureNode infrastructureNode = new InfrastructureNodeParser().parse(getContext(DeploymentNodeDslContext.class), tokens.withoutContextStartToken());
@@ -541,7 +562,7 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
                         }
 
                         registerIdentifier(identifier, infrastructureNode);
-                        parserListener.onParsedModelElement(file, lineNumber, identifier, infrastructureNode);
+                        parserListener.onParsedModelElement(dslFile, lineNumber, identifier, infrastructureNode);
 
                     } else if (SOFTWARE_SYSTEM_INSTANCE_TOKEN.equalsIgnoreCase(firstToken) && inContext(DeploymentNodeDslContext.class)) {
                         SoftwareSystemInstance softwareSystemInstance = new SoftwareSystemInstanceParser().parse(getContext(DeploymentNodeDslContext.class), tokens.withoutContextStartToken());
@@ -551,7 +572,7 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
                         }
 
                         registerIdentifier(identifier, softwareSystemInstance);
-                        parserListener.onParsedModelElement(file, lineNumber, identifier, softwareSystemInstance);
+                        parserListener.onParsedModelElement(dslFile, lineNumber, identifier, softwareSystemInstance);
                         
                     } else if (CONTAINER_INSTANCE_TOKEN.equalsIgnoreCase(firstToken) && inContext(DeploymentNodeDslContext.class)) {
                         ContainerInstance containerInstance = new ContainerInstanceParser().parse(getContext(DeploymentNodeDslContext.class), tokens.withoutContextStartToken());
@@ -561,51 +582,51 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
                         }
 
                         registerIdentifier(identifier, containerInstance);
-                        parserListener.onParsedModelElement(file, lineNumber, identifier, containerInstance);
+                        parserListener.onParsedModelElement(dslFile, lineNumber, identifier, containerInstance);
 
                     } else if (HEALTH_CHECK_TOKEN.equalsIgnoreCase(firstToken) && inContext(StaticStructureElementInstanceDslContext.class)) {
                         new HealthCheckParser().parse(getContext(StaticStructureElementInstanceDslContext.class), tokens.withoutContextStartToken());
                     } else if (CUSTOM_VIEW_TOKEN.equalsIgnoreCase(firstToken) && inContext(ViewsDslContext.class)) {
                         CustomView view = new CustomViewParser().parse(getContext(), tokens.withoutContextStartToken());
                         startContext(new CustomViewDslContext(view));
-                        parserListener.onParsedView(file, lineNumber, view);
+                        parserListener.onParsedView(dslFile, lineNumber, view);
 
                     } else if (SYSTEM_LANDSCAPE_VIEW_TOKEN.equalsIgnoreCase(firstToken) && inContext(ViewsDslContext.class)) {
                         SystemLandscapeView view = new SystemLandscapeViewParser().parse(getContext(), tokens.withoutContextStartToken());
                         startContext(new SystemLandscapeViewDslContext(view));
-                        parserListener.onParsedView(file, lineNumber, view);
+                        parserListener.onParsedView(dslFile, lineNumber, view);
 
                     } else if (SYSTEM_CONTEXT_VIEW_TOKEN.equalsIgnoreCase(firstToken) && inContext(ViewsDslContext.class)) {
                         SystemContextView view = new SystemContextViewParser().parse(getContext(), tokens.withoutContextStartToken());
                         startContext(new SystemContextViewDslContext(view));
-                        parserListener.onParsedView(file, lineNumber, view);
+                        parserListener.onParsedView(dslFile, lineNumber, view);
 
                     } else if (CONTAINER_VIEW_TOKEN.equalsIgnoreCase(firstToken) && inContext(ViewsDslContext.class)) {
                         ContainerView view = new ContainerViewParser().parse(getContext(), tokens.withoutContextStartToken());
                         startContext(new ContainerViewDslContext(view));
-                        parserListener.onParsedView(file, lineNumber, view);
+                        parserListener.onParsedView(dslFile, lineNumber, view);
 
                     } else if (COMPONENT_VIEW_TOKEN.equalsIgnoreCase(firstToken) && inContext(ViewsDslContext.class)) {
                         ComponentView view = new ComponentViewParser().parse(getContext(), tokens.withoutContextStartToken());
                         startContext(new ComponentViewDslContext(view));
-                        parserListener.onParsedView(file, lineNumber, view);
+                        parserListener.onParsedView(dslFile, lineNumber, view);
 
                     } else if (DYNAMIC_VIEW_TOKEN.equalsIgnoreCase(firstToken) && inContext(ViewsDslContext.class)) {
                         DynamicView view = new DynamicViewParser().parse(getContext(), tokens.withoutContextStartToken());
                         startContext(new DynamicViewDslContext(view));
-                        parserListener.onParsedView(file, lineNumber, view);
+                        parserListener.onParsedView(dslFile, lineNumber, view);
 
                     } else if (DEPLOYMENT_VIEW_TOKEN.equalsIgnoreCase(firstToken) && inContext(ViewsDslContext.class)) {
                         DeploymentView view = new DeploymentViewParser().parse(getContext(), tokens.withoutContextStartToken());
                         startContext(new DeploymentViewDslContext(view));
-                        parserListener.onParsedView(file, lineNumber, view);
+                        parserListener.onParsedView(dslFile, lineNumber, view);
 
                     } else if (FILTERED_VIEW_TOKEN.equalsIgnoreCase(firstToken) && inContext(ViewsDslContext.class)) {
                         new FilteredViewParser().parse(getContext(), tokens);
 
                     } else if (tokens.size() > 2 && RELATIONSHIP_TOKEN.equals(tokens.get(1)) && inContext(DynamicViewDslContext.class)) {
-                        Relationship relationship = new DynamicViewContentParser().parseRelationship(getContext(DynamicViewDslContext.class), tokens);
-                        parserListener.onParsedRelationShip(file, lineNumber, identifier, relationship);
+                        new DynamicViewContentParser().parseRelationship(getContext(DynamicViewDslContext.class), tokens);
+                        //parserListener.onParsedRelationShip(dslFile, lineNumber, identifier, relationship);
 
                     } else if (DslContext.CONTEXT_START_TOKEN.equalsIgnoreCase(firstToken) && inContext(DynamicViewDslContext.class)) {
                         startContext(new DynamicViewParallelSequenceDslContext(getContext(DynamicViewDslContext.class)));
@@ -691,6 +712,9 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
                     } else if (VIEW_DESCRIPTION_TOKEN.equalsIgnoreCase(firstToken) && inContext(DeploymentViewDslContext.class)) {
                         new ViewParser().parseDescription(getContext(DeploymentViewDslContext.class), tokens);
 
+                    } else if (inContext(DynamicViewDslContext.class)) {
+                        new DynamicViewContentParser().parseRelationship(getContext(DynamicViewDslContext.class), tokens);
+
                     } else if (THEME_TOKEN.equalsIgnoreCase(firstToken) && inContext(ViewsDslContext.class)) {
                         new ThemeParser().parseTheme(getContext(), tokens);
 
@@ -737,7 +761,7 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
                         if (!restricted || tokens.get(1).startsWith("https://")) {
                             IncludedDslContext context = new IncludedDslContext(dslFile);
                             new IncludeParser().parse(context, tokens);
-                            parserListener.onInclude(file, lineNumber, context.getFile(), tokens.get(1));
+                            parserListener.onInclude(dslFile, lineNumber, context.getFile(), tokens.get(1));
                             parse(context.getLines(), context.getFile());
                             includeInDslSourceLines = false;
                         }